Optimized search function and changed addWatched function to add faster
Coded search function such that now it is able to fetch all possible results for a search query & after the change in the database structure(i.e. removal of cast_url), changed addWatched function such that it directly takes data from search results and adds it to the watch list(i.e. it does not load to fetch info and, directly adds to the database), also tweaked the JS code and set the window.location equal to window.history.back(). So now you have the ability to add any number of movies at once from the search results, as it should be.